% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{aggregator}
\alias{aggregator}
\title{Cartesian product of two ralgets}
\usage{
aggregator(edges)
}
\arguments{
\item{x}{ralget graph}

\item{y}{ralget graph}
}
\description{
Cartesian product of two ralgets
}
