% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.R
\name{b}
\alias{b}
\title{A shorthand to represent coefficient in a data generating process.}
\usage{
b(m)
}
\arguments{
\item{m}{coefficient value}
}
\description{
A shorthand to represent coefficient in a data generating process.
}
