% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{connect}
\alias{connect}
\title{Connect two ralget graphs - original}
\usage{
connect(v1, v2)
}
\arguments{
\item{v1}{a ralget graph}

\item{v2}{a ralget graph}
}
\value{
ralget
}
\description{
Connect two ralget graphs - original
}
