% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_causal.R
\name{const_map}
\alias{const_map}
\title{Set value for variable in DAG}
\usage{
const_map(f, val)
}
\arguments{
\item{f}{a function}

\item{val}{a constant value to replace return values of a function}
}
\description{
Set value for variable in DAG
}
