% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_work_times.R
\name{create_work_times}
\alias{create_work_times}
\title{Create ralget edge}
\usage{
create_work_times(n, interval, start_time = "2021-10-03-9-00")
}
\arguments{
\item{...}{objects stored as vertex attributes.}
}
\description{
Create ralget edge
}
