% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_period.R
\name{cycle_period}
\alias{cycle_period}
\title{Create ralget edge}
\usage{
cycle_period(p_list, time_tracker)
}
\arguments{
\item{...}{objects stored as vertex attributes.}
}
\description{
Create ralget edge
}
