% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{d}
\alias{d}
\title{Specify a data generating process}
\usage{
d(x, env = global_env(), ..., arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{ralget}

\item{env}{assignment expressions for node in DAG}
}
\description{
Specify a data generating process
}
