% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_causal.R
\name{do_causal}
\alias{do_causal}
\title{Set value for variable in DAG}
\usage{
do_causal(graph, ...)
}
\arguments{
\item{g}{ralget}

\item{p}{assignment expressions for node in DAG}
}
\description{
Set value for variable in DAG
}
