% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a graph where nodes and edges store functions.}
\usage{
evaluate(g)
}
\arguments{
\item{g}{ralget}
}
\description{
Evaluate a graph where nodes and edges store functions.
}
\examples{
\dontrun{

 product <- function(...){list(...) \%>\% reduce(`*`)}
 const <- function(x){function(...){x}}
 
 S01 <-   v("S01",   .func = const(10000))
 I01 <-   v("I01",   .func = const(10))
 BIS01 <- v("BIS01", .func = product)
 
 S02 <-   v("S02",   .func = sum)
 I02 <-   v("I02",   .func = sum)
 
 SIR <- 
 S01    * ( e(.func = function(.value){.value * .5} ) * BIS01  + 
            e(.func = function(.value){.value * 1} )  * S02 ) +
 
 I01    * ( e(.func = function(.value){.value * 1} ) * BIS01  +
            e(.func = function(.value){.value * 1}) * I02 )  +
 
 BIS01  *  (e(.func = function(.value){.value * 1} ) * S02  +
            e(.func = function(.value){.value * -2}) *  I02 ) 
 
 evaluate(SIR)
}
}
