% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_execute}
\alias{evaluate_execute}
\title{Evaluate a graph where nodes and edges store functions.2}
\usage{
evaluate_execute(g)
}
\arguments{
\item{g}{ralget}
}
\description{
Evaluate a graph where nodes and edges store functions.2
}
