% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_prepare}
\alias{evaluate_prepare}
\title{Evaluate a graph where nodes and edges store functions.1}
\usage{
evaluate_prepare(g)
}
\arguments{
\item{g}{ralget}
}
\description{
Evaluate a graph where nodes and edges store functions.1
}
