% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{extract_edgelist_n}
\alias{extract_edgelist_n}
\title{Multiply two ralget graphs}
\usage{
extract_edgelist_n(graph, n, side)
}
\arguments{
\item{graph}{a ralget graph}

\item{n}{a ralget graph}

\item{side}{a ralget graph}
}
\value{
ralget
}
\description{
Multiply two ralget graphs
}
