% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_t.R
\name{increment_t}
\alias{increment_t}
\title{Set value for variable in DAG}
\usage{
increment_t(g, t = 1)
}
\arguments{
\item{g}{evaluated ralget graph with nodes ending in numbers that represent time periods}

\item{t}{number of periods to move the dag forward}
}
\description{
Set value for variable in DAG
}
