% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instantiate_parent.R
\name{instantiate_parent}
\alias{instantiate_parent}
\title{Create ralget edge}
\usage{
instantiate_parent(task)
}
\arguments{
\item{...}{objects stored as vertex attributes.}
}
\description{
Create ralget edge
}
