% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_package_silent.r
\name{load_package_silent}
\alias{load_package_silent}
\title{Create ralget edge}
\usage{
load_package_silent(package_name, verbose_error = FALSE)
}
\arguments{
\item{...}{objects stored as vertex attributes.}
}
\description{
Create ralget edge
}
