% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Multiply two ralget graphs - original}
\usage{
multiply(v1, v2)
}
\arguments{
\item{v1}{a ralget graph}

\item{v2}{a ralget graph}
}
\value{
ralget
}
\description{
Multiply two ralget graphs - original
}
