% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{overlay}
\alias{overlay}
\title{Add two ralget graphs - original}
\usage{
overlay(v1, v2)
}
\arguments{
\item{v1}{a ralget graph}

\item{v2}{a ralget graph}
}
\value{
ralget
}
\description{
Add two ralget graphs - original
}
