% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_executed_plan.R
\name{plot_executed_plan}
\alias{plot_executed_plan}
\title{Create ralget edge}
\usage{
plot_executed_plan(comined)
}
\arguments{
\item{...}{objects stored as vertex attributes.}
}
\description{
Create ralget edge
}
