% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_joint_distributions.R
\name{plot_joint_distributions}
\alias{plot_joint_distributions}
\title{Simulate data from a dag}
\usage{
plot_joint_distributions(df)
}
\arguments{
\item{df}{plot joint distributions of a simulated df}
}
\description{
Simulate data from a dag
}
