% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{+.ralget_edge}
\alias{+.ralget_edge}
\title{Add two ralget graphs}
\usage{
\method{+}{ralget_edge}(v1, v2)
}
\arguments{
\item{v1}{a ralget graph}

\item{v2}{a ralget graph}
}
\value{
ralget
}
\description{
Add two ralget graphs
}
