% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ralget_graph_join.R
\name{ralget_graph_join}
\alias{ralget_graph_join}
\title{Create ralget edge}
\usage{
ralget_graph_join(vl, vr)
}
\arguments{
\item{...}{objects stored as vertex attributes.}
}
\description{
Create ralget edge
}
