% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ralget_sim.R
\name{render}
\alias{render}
\title{concat to vectors  for data sim}
\usage{
render(x, y, n)
}
\arguments{
\item{x}{vertices}

\item{y}{vertices}

\item{n}{vertices}
}
\description{
concat to vectors  for data sim
}
