% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{separate_edges}
\alias{separate_edges}
\title{Multiply two ralget graphs}
\usage{
separate_edges(current_graph, side)
}
\arguments{
\item{current_graph}{a ralget graph}

\item{side}{a ralget graph}
}
\value{
ralget
}
\description{
Multiply two ralget graphs
}
