% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate data from a dag}
\usage{
simulate(g, label, seed = NULL)
}
\arguments{
\item{g}{a ralget DAG}

\item{label}{the label associated with simulated data.}
}
\description{
Simulate data from a dag
}
