% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{*.ralget}
\alias{*.ralget}
\title{Multiply two ralget graphs}
\usage{
\method{*}{ralget}(v1, v2)
}
\arguments{
\item{v1}{a ralget graph}

\item{v2}{a ralget graph}
}
\value{
ralget
}
\description{
Multiply two ralget graphs
}
