% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_status.R
\name{update_status}
\alias{update_status}
\title{Create ralget edge}
\usage{
update_status(formed_plan_01)
}
\arguments{
\item{...}{objects stored as vertex attributes.}
}
\description{
Create ralget edge
}
