% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian_product.R
\name{cartesian_product}
\alias{cartesian_product}
\title{Cartesian product of two ralgets}
\usage{
cartesian_product(x, y)
}
\arguments{
\item{x}{ralget graph}

\item{y}{ralget graph}
}
\description{
Cartesian product of two ralgets
}
