% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{d}
\alias{d}
\title{Specify a data generating process}
\usage{
d(x, env = caller_env())
}
\arguments{
\item{...}{a data generating process as an anonymous equation}
}
\description{
Specify a data generating process
}
